Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0



